+++ relay-event-normalization/src/timestamp.rs
@@ -1,15 +1,15 @@
 use relay_event_schema::processor::{
     ProcessValue, ProcessingAction, ProcessingResult, ProcessingState, Processor,
 };
-use relay_event_schema::protocol::{Event, Span};
+use relay_event_schema::protocol::{Breadcrumb, Event, Span};
 use relay_protocol::{Error, Meta};
 
 /// Ensures an event's timestamps are not stale.
 ///
 /// Stale timestamps are those that happened before January 1, 1970 UTC. The
 /// processor validates the start and end timestamps of an event and returns an
-/// error if any of these are stale. Additionally, spans with stale timestamps
-/// are removed from the event.
+/// error if any of these are stale. Additionally, spans and breadcrumbs with
+/// stale timestamps are removed from the event.
 ///
 /// The processor checks the timestamps individually and it's not responsible
 /// for decisions that relate timestamps together, including but not limited to:
@@ -71,12 +71,31 @@ impl Processor for TimestampProcessor {
 
         Ok(())
     }
+
+    fn process_breadcrumb(
+        &mut self,
+        breadcrumb: &mut Breadcrumb,
+        meta: &mut Meta,
+        _: &ProcessingState<'_>,
+    ) -> ProcessingResult where {
+        if let Some(timestamp) = breadcrumb.timestamp.value() {
+            if timestamp.into_inner().timestamp_millis() < 0 {
+                meta.add_error(Error::invalid(format!(
+                    "timestamp is too stale: {}",
+                    timestamp
+                )));
+                return Err(ProcessingAction::DeleteValueHard);
+            }
+        }
+
+        Ok(())
+    }
 }
 
 #[cfg(test)]
 mod tests {
     use relay_event_schema::processor::{process_value, ProcessingState};
-    use relay_event_schema::protocol::{Event, Span};
+    use relay_event_schema::protocol::{Breadcrumb, Event, Span};
     use relay_protocol::{assert_annotated_snapshot, get_value, Annotated};
 
     use crate::timestamp::TimestampProcessor;
@@ -226,6 +245,54 @@ mod tests {
                   "invalid_data",
                   {
                     "reason": "start_timestamp is too stale: 1969-12-31 23:59:59 UTC"
+                  }
+                ]
+              ]
+            }
+          }
+        }
+        "###);
+    }
+
+    #[test]
+    fn test_accept_recent_breadcrumb() {
+        let json = r#"{
+      "timestamp": 1
+    }"#;
+        let mut breadcrumb = Annotated::<Breadcrumb>::from_json(json).unwrap();
+        assert!(process_value(
+            &mut breadcrumb,
+            &mut TimestampProcessor,
+            ProcessingState::root()
+        )
+        .is_ok());
+        assert_eq!(
+            get_value!(breadcrumb.timestamp!).into_inner().timestamp(),
+            1
+        );
+    }
+
+    #[test]
+    fn test_reject_stale_breadcrumb() {
+        let json = r#"{
+      "timestamp": -1
+    }"#;
+        let mut breadcrumb = Annotated::<Breadcrumb>::from_json(json).unwrap();
+        assert!(process_value(
+            &mut breadcrumb,
+            &mut TimestampProcessor,
+            ProcessingState::root()
+        )
+        .is_ok());
+        assert_annotated_snapshot!(&breadcrumb, @r###"
+        {
+          "_meta": {
+            "": {
+              "err": [
+                [
+                  "invalid_data",
+                  {
+                    "reason": "timestamp is too stale: 1969-12-31 23:59:59 UTC"
                   }
                 ]
               ]